ai_interview_app/requirements.txt
annotated-types==0.7.0
anyio==4.5.2
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
distro==1.9.0
exceptiongroup==1.2.2
gTTS==2.5.4
h11==0.14.0
httpcore==1.0.8
httpx==0.28.1
idna==3.10
jiter==0.9.0
pygame==2.6.1
pydantic==2.10.6
pydantic-core==2.27.2
python-dotenv==1.0.1
requests==2.32.3
sniffio==1.3.1
SpeechRecognition==3.10.4
tqdm==4.67.1
typing-extensions==4.13.2
urllib3==2.2.3

# Backend + LangChain
fastapi==0.115.2
uvicorn==0.32.0
jinja2==3.1.4
langchain==0.2.14
langchain-community==0.2.11
groq==0.31.1
python-multipart

